// src/app/models/complaint.ts
export interface Complaint {    

    InternalId?: any;

    ComplaintId?: number;
    SubscriptionNo?: string;
    Description?: string;
    Status?: string;
    Comments?: string;
    CreatedDate?: any;
    ModifiedDate?: any;

}

// public class Complaint
//     {
//         [BsonId]
//         // standard BSonId generated by MongoDb
//         public ObjectId InternalId { get; set; }

//         public int ComplaintId { get; set; }
//         public string SubscriptionNo { get; set; }
//         public string Description { get; set; }
//         public string Status { get; set; }
//         public string Comments { get; set; }

//         public DateTime CreatedDate { get; set; }
//         public DateTime ModifiedDate { get; set; }
//     }